+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Unreleased
 * BREAKING CHANGE: ``market_state`` dict will now always include ``tse_value`` and ``tse_index_change_percent`` keys. The value will be ``None`` if not available.
 * BREAKING CHANGE: Use ``ins_code`` as the index for ``price`` and ``new_prices`` DataFrames in ``market_watch_init`` and ``market_watch_plus``. Previously, the index was a MultiIndex of ``['ins_code', 'isin', 'l18', 'l30']``.
 * Fixed a bug in parsing ``market_state`` during pre-market.
+* Updated dataset
 
 v0.50.0
 -------